requirements.txt
# Requirements for Medical Information Chatbot System

# Core Dependencies
haystack-ai>=2.0.0
sentence-transformers>=2.2.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: For better performance
torch>=2.0.0  # Required by sentence-transformers, install CPU or GPU version as needed

# Existing dependencies (from original project)
# No additional ML frameworks needed - lightweight design

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
